+++ packages/client/CHANGELOG.md
@@ -7,6 +7,10 @@
 - `polkadot-api/smoldot/node-worker` to create a smoldot worker in NodeJS.
 - `polkadot-api/smoldot/from-node-worker` to create a client from a node smoldot worker.
 
+### Fixed
+
+- Fixed correct ESM export for React Native.
+
 ## 0.12.1 - 2024-07-30
 
 ### Fixed