+++ packages/client/CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Descriptors entry points when transaction and events have the same checksum.
 - Fixed correct ESM export for React Native.
+- `chains`: Update `lightSyncState`
 
 ## 0.12.1 - 2024-07-30