+++ packages/client/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.13.0 - 2024-08-12
+
 ### Added
 
 - `polkadot-api/smoldot/node-worker` to create a smoldot worker in NodeJS.